+++ wtfix/apps/utils.py
@@ -14,10 +14,11 @@
 #
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+from unsync import unsync
 
 from wtfix.apps.base import BaseApp
 from wtfix.conf import settings
-
+from wtfix.message.message import FIXMessage
 
 logger = settings.logger
 
@@ -29,10 +30,11 @@ class OutboundLoggingApp(BaseApp):
 
     name = "outbound_logger"
 
-    def on_send(self, message):
+    @unsync
+    async def on_send(self, message: FIXMessage) -> FIXMessage:
         logger.info(f" --> {message:t}")
 
-        return message
+        return super().on_send(message)
 
 
 class InboundLoggingApp(BaseApp):
@@ -42,7 +44,8 @@ class InboundLoggingApp(BaseApp):
 
     name = "inbound_logger"
 
-    def on_receive(self, message):
+    @unsync
+    async def on_receive(self, message: FIXMessage) -> FIXMessage:
         logger.info(f" <-- {message:t}")
 
-        return message
+        return await super().on_receive(message)